Consolidate compaitibility tests
